Added login and register pages and functionality. after login will be re directed to home page(currently empty) tried adding a light and dark mode button, but has issues
want to add persistency to the theme if i have time so once the user changes page it doesnt revert back to light theme.

Need to work on home page to add functionality for getting friends and channels from the back end.

Thinking of adding a time stamp on messages when storing them so when sorting the order of messages will be correct.